# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Family Tree API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/family-trees/{tree_id}/persons": {
      "post": {
        "tags": [
          "Persons"
        ],
        "summary": "Create Person",
        "operationId": "create_person_family_trees__tree_id__persons_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Persons"
        ],
        "summary": "Update Person",
        "operationId": "update_person_family_trees__tree_id__persons_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/family-trees/{tree_id}/persons/{person_id}": {
      "delete": {
        "tags": [
          "Persons"
        ],
        "summary": "Delete Person",
        "operationId": "delete_person_family_trees__tree_id__persons__person_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Get Person",
        "operationId": "get_person_family_trees__tree_id__persons__person_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonGetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/family-trees/{tree_id}/persons/{from_person_id}/relation/{to_person_id}/alternatives": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Get Alternative Relationship Paths",
        "operationId": "get_alternative_relationship_paths_family_trees__tree_id__persons__from_person_id__relation__to_person_id__alternatives_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "from_person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "From Person Id"
            }
          },
          {
            "name": "to_person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "To Person Id"
            }
          },
          {
            "name": "male_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Male Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosestRelationshipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/family-trees/{tree_id}/persons/{from_person_id}/relation/{to_person_id}": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Get Closest Relationship",
        "operationId": "get_closest_relationship_family_trees__tree_id__persons__from_person_id__relation__to_person_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "from_person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "From Person Id"
            }
          },
          {
            "name": "to_person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "To Person Id"
            }
          },
          {
            "name": "male_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Male Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosestRelationshipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/family-trees/{tree_id}/persons/list": {
      "post": {
        "tags": [
          "Persons"
        ],
        "summary": "Get Person List By Filter",
        "operationId": "get_person_list_by_filter_family_trees__tree_id__persons_list_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterPersonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_PersonModel_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClosestRelationshipResponse": {
        "properties": {
          "from_person_id": {
            "type": "string",
            "format": "uuid",
            "title": "From Person Id"
          },
          "to_person_id": {
            "type": "string",
            "format": "uuid",
            "title": "To Person Id"
          },
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "distance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance"
          },
          "path_person_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Path Person Ids"
          },
          "relationship_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Relationship Types"
          },
          "paths": {
            "items": {
              "$ref": "#/components/schemas/RelationshipPathItemResponse"
            },
            "type": "array",
            "title": "Paths"
          }
        },
        "type": "object",
        "required": [
          "from_person_id",
          "to_person_id",
          "found"
        ],
        "title": "ClosestRelationshipResponse"
      },
...
      "FilterPersonRequest": {
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequestParams"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonFilterRequestData"
              },
              {
                "type": "null"
              }
            ]
          },
          "sort": {
            "$ref": "#/components/schemas/SortRequestParams_PersonSortField_"
          }
        },
        "type": "object",
        "title": "FilterPersonRequest"
      },
...
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female"
        ],
        "title": "Gender"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_PersonModel_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonModel"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedResponse[PersonModel]"
      },
...
      "ParentLinkRequest": {
        "properties": {
          "parent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Parent Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/ParentRelationshipType",
            "default": "biological"
          }
        },
        "type": "object",
        "required": [
          "parent_id"
        ],
        "title": "ParentLinkRequest"
      },
      "ParentRelationshipType": {
        "type": "string",
        "enum": [
          "biological",
          "adoptive",
          "step"
        ],
        "title": "ParentRelationshipType"
      },
...
      "PersonCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Person full name"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "death_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Date"
          },
          "family_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "birth_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Place"
          },
          "death_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Place"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/ParentLinkRequest"
            },
            "type": "array",
            "title": "Parents"
          },
          "marriage_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marriage Id"
          },
          "photo_object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Object Key"
          }
        },
        "type": "object",
        "required": [
          "name",
          "gender"
        ],
        "title": "PersonCreateRequest",
        "example": {
          "birth_date": "2000-11-21",
          "gender": "male",
          "name": "arash",
          "parents": [
            {
              "parent_id": "00000000-0000-0000-0000-000000000001",
              "relationship_type": "biological"
            },
            {
              "parent_id": "00000000-0000-0000-0000-000000000002",
              "relationship_type": "biological"
            }
          ]
        }
      },
      "PersonCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "death_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Date"
          },
          "family_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "birth_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Place"
          },
          "death_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Place"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/ParentLinkRequest"
            },
            "type": "array",
            "title": "Parents"
          },
          "marriage_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marriage Id"
          },
          "photo_object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Object Key"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "gender",
          "birth_date"
        ],
        "title": "PersonCreateResponse",
        "example": {
          "birth_date": "2000-11-21",
          "gender": "male",
          "name": "arash",
          "parents": [
            {
              "parent_id": "00000000-0000-0000-0000-000000000001",
              "relationship_type": "biological"
            }
          ]
        }
      },
...
      "PersonGetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "death_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Date"
          },
          "family_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "birth_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Place"
          },
          "death_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Place"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/ParentLinkRequest"
            },
            "type": "array",
            "title": "Parents"
          },
          "marriage_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marriage Id"
          },
          "photo_object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Object Key"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "gender",
          "birth_date"
        ],
        "title": "PersonGetResponse",
        "example": {
          "birth_date": "1996-07-31",
          "gender": "male",
          "name": "Ali"
        }
      },
...
      "PersonUpdateRequest": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/_PersonUpdateDateRequest"
          },
          "where": {
            "$ref": "#/components/schemas/_PersonUpdateWhereRequest"
          }
        },
        "type": "object",
        "required": [
          "data",
          "where"
        ],
        "title": "PersonUpdateRequest"
      },
      "PersonUpdateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "death_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Date"
          },
          "family_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "birth_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Place"
          },
          "death_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Place"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/ParentLinkRequest"
            },
            "type": "array",
            "title": "Parents"
          },
          "marriage_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marriage Id"
          },
          "photo_object_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Object Key"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "gender",
          "birth_date"
        ],
        "title": "PersonUpdateResponse"
      },
...
      "ResultResponse": {
        "properties": {
          "result": {
            "type": "string",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "ResultResponse"
      },
...
      "_PersonUpdateWhereRequest": {
        "properties": {
          "person_id": {
            "type": "string",
            "format": "uuid",
            "title": "Person Id"
          }
        },
        "type": "object",
        "required": [
          "person_id"
        ],
        "title": "_PersonUpdateWhereRequest"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}